fix: fix account deletion bug when last sign in was more than five minutes ago

+++ src/app/pages/profile/profile.component.ts
@@ -138,15 +138,19 @@ export class ProfileComponent implements OnInit {
     dialogRef.afterClosed().subscribe((confirmed) => {
       if (confirmed) {
         this.authService.forceReauthentication().then(() => {
-          this.userService.deleteUser(userId).then(() => {
-            this.authService.deleteAccount().then(() => {
-              this.router.navigate(['/']);
+          if(this.auth.currentUser) {
+            this.userService.deleteUser(userId).then(() => {
+              this.authService.deleteAccount().then(() => {
+                this.router.navigate(['/']);
+              }).catch((error) => {
+                console.error('Failed to delete firebase account:', error);
+              });
             }).catch((error) => {
-              console.error('Failed to delete firebase account:', error);
+              console.error('Failed to delete user', error);
             });
-          }).catch((error) => {
-            console.error('Failed to delete user', error);
-          });
+          } else {
+            this.router.navigate(['/login']);
+          }
         }).catch((error) => {
           console.error('Failed to reauthenticate:', error);
         });
